refactor: implement buildAnswersPayload function for cleaner answer handling in FormDetailPage
